streamlit/requirements.txt
streamlit
pytesseract
paddleocr

pypdfium2
pillow
transformers
torch
opencv-python-headless
numpy
